requirements.txt
fastapi
uvicorn
python-multipart
pytesseract
Pillow
transformers
torch
